+++ lib/extend/highlight/index.js
@@ -1,0 +1,11 @@
+'use strict';
+
+module.exports = hexo => {
+
+  const { loadUtil } = require('../../utils');
+  const load = loadUtil(hexo);
+  if (!hexo.config.disable_marked_highlight) {
+    hexo.config.highlight.enable = false;
+    load('hexo-filter-marked-highlight');
+  }
+};
