+++ lib/extend/index.js
@@ -5,5 +5,6 @@ module.exports = hexo => {
   require('./medium-zoom')(hexo);
   require('./reward')(hexo);
   require('./marked')(hexo);
+  require('./highlight')(hexo);
   require('./external-link')(hexo);
 };
