+++ lib/load.js
@@ -1,16 +1,6 @@
 'use strict';
 
-const { join } = require('path');
-
-const loadUtil = hexo => {
-  const packagePath = join(hexo.base_dir, 'package.json');
-  const deps = require(packagePath).dependencies;
-  return name => {
-    if (!deps[name]) {
-      hexo.loadPlugin(hexo.resolvePlugin(name));
-    }
-  };
-};
+const { loadUtil } = require('./utils');
 
 module.exports = hexo => {
   if (hexo.cake_scripts_loaded) {
@@ -31,8 +21,6 @@ module.exports = hexo => {
   const load = loadUtil(hexo);
   load('hexo-fontawesome');
   load('hexo-renderer-ejs');
-  hexo.config.highlight.enable = false;
-  load('hexo-filter-marked-highlight');
   const { filter } = hexo.extend;
   filter.register('after_init', () => {
     load('@jiangtj/hexo-renderer-stylus');
