+++ lib/utils.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const { mergeWith } = require('lodash');
+const { join } = require('path');
 
 module.exports = {
   htmlTag: function(tag, attrs, text) {
@@ -20,5 +21,15 @@ module.exports = {
         return srcValue;
       }
     });
+  },
+  deps    : null,
+  loadUtil: hexo => {
+    const packagePath = join(hexo.base_dir, 'package.json');
+    if (!this.deps) this.deps = require(packagePath).dependencies;
+    return name => {
+      if (!this.deps[name]) {
+        hexo.loadPlugin(hexo.resolvePlugin(name));
+      }
+    };
   }
 };
